file read ok

+++ defs.h
@@ -7,8 +7,8 @@
 /**/
 void A0(); 
 
-/*Returns 1 if there's a route between the two airports, otherwise returns 0 */
-int B0(Graph *, int, int); 
+/*Returns 1 if there's a route between the two airports, otherwise returns 0*/
+double B0(Graph *, int, int); 
 
 /**/
 void C0();
